+++ anchor/programs/games/src/errors.rs
@@ -11,6 +11,9 @@ pub enum GameErrorCode {
     #[msg("Name must be between 3 and 32 characters")]
     InvalidName,
 
+    #[msg("Game is not active")]
+    GameInactive,
+
     #[msg("Game is already ended")]
     GameEnded,
 
@@ -20,7 +23,7 @@ pub enum GameErrorCode {
 
 #[error_code]
 pub enum GameModeErrorCode {
-    #[msg("Slots must be between 1 and 16")]
+    #[msg("Slots must be between 1 and 10")]
     InvalidSlots,
 
     #[msg("Digits must be between 1 and 8")]
@@ -52,4 +55,22 @@ pub enum BountyErrorCode {
 
     #[msg("Total vault reward is uncollectible")]
     UncollectibleReward,
+}
+
+#[error_code]
+pub enum PlayerErrorCode {
+    #[msg("Game provided does not match the game mode")]
+    GameMismatch,
+
+    #[msg("Mode provided does not match the bounty task")]
+    TaskMismatch,
+}
+
+#[error_code]
+pub enum RoundErrorCode {
+    #[msg("Invalid seed")]
+    InvalidSeed,
+
+    #[msg("Invalid choice")]
+    InvalidChoice,
 }