+++ web/hooks/games/index.ts
@@ -3,4 +3,6 @@ export * from './use-games-program';
 export * from './game';
 export * from './game-mode';
 export * from './treasure';
+export * from './escrow';
 export * from './bounty';
+export * from './player';